+++ src/github.com/cthit/goldapps/admin/password.go
