+++ src/github.com/cthit/goldapps/admin/scope.go
