+++ src/github.com/cthit/goldapps/admin/service.go
@@ -1,7 +1,7 @@
 package admin
 
 import (
-	"github.com/cthit/goldapps"
+	"../../goldapps"
 
 	"google.golang.org/api/admin/directory/v1" // Imports as admin
 	"google.golang.org/api/gmail/v1"           // Imports as gmail
