+++ src/github.com/cthit/goldapps/cmd/config.go
