+++ src/github.com/cthit/goldapps/cmd/flags.go
