+++ src/github.com/cthit/goldapps/cmd/util.go
