+++ src/github.com/cthit/goldapps/group_action.go
