+++ src/github.com/cthit/goldapps/util.go
